+++ matrix/admin.go
@@ -1,6 +1,9 @@
 package matrix
 
 import (
+	"errors"
+	"fmt"
+	"net/url"
 	"time"
 
 	"github.com/turt2live/matrix-media-repo/common/rcontext"
@@ -16,12 +19,12 @@ func IsUserAdmin(ctx rcontext.RequestContext, serverName string, accessToken str
 	replyError = cb.CallContext(ctx, func() error {
 
 		response := &whoisResponse{}
-		path := "/_matrix/client/unstable/admin/whois/"
+		path := fmt.Sprintf("/_matrix/client/v3/admin/whois/%s", url.PathEscape(fakeUser))
 		if hs.AdminApiKind == "synapse" {
-			path = "/_synapse/admin/v1/whois/"
+			path = fmt.Sprintf("/_synapse/admin/v1/whois/%s", url.PathEscape(fakeUser))
 		}
-		url := util.MakeUrl(hs.ClientServerApi, path, fakeUser)
-		err := doRequest(ctx, "GET", url, nil, response, accessToken, ipAddr)
+		urlStr := util.MakeUrl(hs.ClientServerApi, path)
+		err := doRequest(ctx, "GET", urlStr, nil, response, accessToken, ipAddr)
 		if err != nil {
 			err, replyError = filterError(err)
 			return err
@@ -40,12 +43,15 @@ func ListMedia(ctx rcontext.RequestContext, serverName string, accessToken strin
 	response := &mediaListResponse{}
 	var replyError error
 	replyError = cb.CallContext(ctx, func() error {
-		path := "/_matrix/client/unstable/admin/room/"
+		path := ""
 		if hs.AdminApiKind == "synapse" {
-			path = "/_synapse/admin/v1/room/"
+			path = fmt.Sprintf("/_synapse/admin/v1/room/%s/media", url.PathEscape(roomId))
 		}
-		url := util.MakeUrl(hs.ClientServerApi, path, roomId, "/media")
-		err := doRequest(ctx, "GET", url, nil, response, accessToken, ipAddr)
+		if path == "" {
+			return errors.New("unable to query media for homeserver: wrong or incompatible adminApiKind")
+		}
+		urlStr := util.MakeUrl(hs.ClientServerApi, path)
+		err := doRequest(ctx, "GET", urlStr, nil, response, accessToken, ipAddr)
 		if err != nil {
 			err, replyError = filterError(err)
 			return err