+++ CHANGELOG.md
@@ -97,6 +97,7 @@ path/server, for example, then you can simply update the path in the config for 
 * Per-user upload quotas now do not allow users to exceed the maximum values, even by 1 byte. Previously, users could exceed the limits by a little bit.
 * Updated to Go 1.19
 * Logs are now less noisy by default.
+* Connected homeservers must support at least Matrix 1.1 on the Client-Server API.
 
 ### Fixed